Updated role of extra params in query keys for datafetching hooks
