+++ packages/react-utils/src/components/FhirSelect/utils.ts
@@ -1,7 +1,12 @@
 import { URLParams } from '@opensrp/server-service';
 import { IBundle } from '@smile-cdr/fhirts/dist/FHIR-R4/interfaces/IBundle';
 import { FHIRServiceClass } from '../../helpers/dataLoaders';
-import { FhirApiFilter } from '../../hooks/useSimpleTabularView';
+
+export interface FhirApiFilter {
+  page: number;
+  pageSize: number;
+  search: string | null;
+}
 
 /**
  * Unified function that gets a list of FHIR resources from a FHIR hapi server
