+++ packages/react-utils/src/hooks/useTabularViewWithLocalSearch.ts
@@ -5,7 +5,6 @@ import { useQuery } from 'react-query';
 import { getConfig } from '@opensrp/pkg-config';
 import { useHistory, useLocation, useRouteMatch } from 'react-router';
 import type { IBundle } from '@smile-cdr/fhirts/dist/FHIR-R4/interfaces/IBundle';
-import { Resource } from '@smile-cdr/fhirts/dist/FHIR-R4/classes/resource';
 import { URLParams } from '@opensrp/server-service';
 import { loadAllResources } from '../helpers/fhir-utils';
 import {
@@ -28,6 +27,7 @@ import {
  * @param resourceType - resource type as endpoint
  * @param extraParams - further custom search param filters during api requests
  * @param matchesSearch -  function that computes whether a resource payload should be matched by search
+ * @param dataTransformer - function to process data after fetch
  */
 export function useTabularViewWithLocalSearch<T extends object>(
   fhirBaseUrl: string,
@@ -59,7 +59,9 @@ export function useTabularViewWithLocalSearch<T extends object>(
   const rQuery = {
     queryKey: [resourceType, extraParams] as TRQuery,
     queryFn,
-    select: (data: IBundle) => dataTransformer(data),
+    select: (data: IBundle) => {
+      return dataTransformer(data);
+    },
     refetchOnWindowFocus: false,
     refetchOnReconnect: false,
   };
